Adjust target test locations and structure
